🚧 (wip) extract state from line charts
