+++ packages/@ourworldindata/grapher/src/chart/ChartInterface.ts
@@ -8,6 +8,7 @@ import {
 import { ColorScale } from "../color/ColorScale"
 import { HorizontalAxis, VerticalAxis } from "../axis/Axis"
 import { HorizontalColorLegendManager } from "../horizontalColorLegend/HorizontalColorLegends"
+
 // The idea of this interface is to try and start reusing more code across our Chart classes and make it easier
 // for a dev to work on a chart type they haven't touched before if they've worked with another that implements
 // this interface.
@@ -19,22 +20,32 @@ export interface ChartSeries {
 
 export type ChartTableTransformer = (inputTable: OwidTable) => OwidTable
 
-export interface ChartInterface {
+/** Interface implemented by all chart state classes */
+export interface ChartState {
     failMessage: string // We require every chart have some fail message(s) to show to the user if something went wrong
 
     inputTable: OwidTable // Points to the OwidTable coming into the chart. All charts have an inputTable. Standardized as part of the interface as a development aid.
     transformedTable: OwidTable // Points to the OwidTable after the chart has transformed the input table. The chart may add a relative transform, for example. Standardized as part of the interface as a development aid.
 
-    colorScale?: ColorScale
-    shouldUseValueBasedColorScheme?: boolean // Opt-out of assigned colors and use a value-based color scheme instead
-
-    seriesStrategy?: SeriesStrategy
     series: readonly ChartSeries[] // This points to the marks that the chart will render. They don't have to be placed yet. Standardized as part of the interface as a development aid.
-    // Todo: should all charts additionally have a placedSeries: ChartPlacedSeries[] getter?
+    seriesStrategy?: SeriesStrategy
 
     transformTable: ChartTableTransformer
     transformTableForDisplay?: ChartTableTransformer
     transformTableForSelection?: ChartTableTransformer
+
+    colorScale?: ColorScale
+
+    /**
+     * Which facet strategies the chart type finds reasonable in its current setting, if any.
+     * Does not necessarily contain FacetStrategy.None -- there are situations where an unfaceted chart doesn't make sense.
+     */
+    availableFacetStrategies?: FacetStrategy[]
+}
+
+/** Interface implemented by all chart component classes */
+export interface ChartInterface {
+    chartState: ChartState
 
     yAxis?: HorizontalAxis | VerticalAxis
     xAxis?: HorizontalAxis | VerticalAxis
@@ -46,8 +57,8 @@ export interface ChartInterface {
     externalLegend?: HorizontalColorLegendManager
 
     /**
-     * Which facet strategies the chart type finds reasonable in its current setting, if any.
-     * Does not necessarily contain FacetStrategy.None -- there are situations where an unfaceted chart doesn't make sense.
+     * Opt-out of assigned colors and use a value-based color scheme instead.
+     * Only relevant for StackedBar charts.
      */
-    availableFacetStrategies?: FacetStrategy[]
+    shouldUseValueBasedColorScheme?: boolean
 }
