+++ packages/@ourworldindata/grapher/src/mapCharts/MapSparkline.tsx
@@ -12,6 +12,7 @@ import {
 } from "@ourworldindata/types"
 import { OwidTable } from "@ourworldindata/core-table"
 import { LineChart } from "../lineCharts/LineChart"
+import { LineChartState } from "../lineCharts/LineChartState"
 import { Bounds, checkIsVeryShortUnit } from "@ourworldindata/utils"
 import { LineChartManager } from "../lineCharts/LineChartConstants"
 import { ColorScale } from "../color/ColorScale.js"
@@ -174,10 +175,14 @@ export class MapSparkline extends React.Component<{
         })
     }
 
+    @computed private get sparklineChartState(): LineChartState {
+        return new LineChartState({ manager: this.sparklineManager })
+    }
+
     @computed private get sparklineChart(): LineChart {
         return new LineChart({
-            manager: this.sparklineManager,
             bounds: this.sparklineBounds,
+            chartState: this.sparklineChartState,
         })
     }
 
@@ -230,8 +235,8 @@ export class MapSparkline extends React.Component<{
                         y2={axisBounds.y}
                     />
                     <LineChart
-                        manager={this.sparklineManager}
                         bounds={this.sparklineBounds}
+                        chartState={this.sparklineChartState}
                     />
                     {maxLabel !== minLabel && (
                         <g className="max axis-label">
